refactor and add flow user
