+++ src/Component/MainBody.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function MainBody() {
+  return <div>MainBody</div>;
+}
+
+export default MainBody;
