+++ src/Component/WhyShouldUse.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function WhyShouldUse() {
+  return (
+    <div>
+    
+    </div>
+  );
+}
+
+export default WhyShouldUse;
